+++ parser/config.go
@@ -164,6 +164,7 @@ var SupportedFunctions = map[string]int{
 	"doublefactorial": 1,
 	"risingfact": 2,
 	"multinomial": -1,
+	"partition": 1,
 	"sumproperdivisors": 1,
 	"amicable": 2,
 	"bigomega":  1,